Move classes from virtual_machine.py to own files.
virtual_machine.py had the following classes:
VirtualMachineInstance
VirtualMachineInstanceMigration
VirtualMachineInstancePreset
VirtualMachineInstanceReplicaSet

each moved to own file.
