+++ ocp_resources/virtual_machine_Instance_replica_set.py
@@ -1,0 +1,20 @@
+from ocp_resources.resource import NamespacedResource
+
+
+class VirtualMachineInstanceReplicaSet(NamespacedResource):
+    """
+    VirtualMachineInstancePreset object.
+    """
+
+    api_group = NamespacedResource.ApiGroup.KUBEVIRT_IO
+
+    def __init__(
+        self, name=None, namespace=None, client=None, teardown=True, yaml_file=None
+    ):
+        super().__init__(
+            name=name,
+            namespace=namespace,
+            client=client,
+            teardown=teardown,
+            yaml_file=yaml_file,
+        )
